HelpersTask717: Factor out file selection
Pre-commit checks:
All checks passed ✅
